requirements.txt
flask
pandas
plotly
dash
numpy
pyserial